requirements.txt
Django==1.11.5
django-rest-framework-mongoengine==3.3.1
mysqlclient==1.3.12
pytz==2017.2
virtualenv==15.1.0
